Add the new getDates() to all models that have timestamps = true
