+++ src/Models/Model.php
@@ -2,6 +2,7 @@
 
 namespace A17\Twill\Models;
 
+use A17\Twill\Models\Behaviors\HasDates;
 use A17\Twill\Models\Behaviors\HasPresenter;
 use A17\Twill\Models\Behaviors\IsTranslatable;
 use Carbon\Carbon;
@@ -14,7 +15,7 @@ use Illuminate\Support\Str;
 
 abstract class Model extends BaseModel implements TaggableInterface
 {
-    use HasPresenter, SoftDeletes, TaggableTrait, IsTranslatable;
+    use HasPresenter, SoftDeletes, TaggableTrait, IsTranslatable, HasDates;
 
     public $timestamps = true;
 
@@ -120,41 +121,5 @@ abstract class Model extends BaseModel implements TaggableInterface
             'taggable_id',
             'tag_id'
         );
-    }
-
-    /**
-     * Get the attributes that should be converted to dates.
-     *
-     * @return array
-     */
-    public function getDates()
-    {
-        $dates = parent::getDates();
-
-        // If it's not Laravel 10, stick to the current implementation
-        if(app()->version() < '10') {
-            return $dates;
-        }
-
-        // If we have a $dates property, get the dates from it
-        if (property_exists($this, 'dates')) {
-            $dates = array_merge($dates, $this->dates);
-        }
-
-        // If we don't have a $casts or it's empty, return the dates we have now
-        if (!property_exists($this, 'casts') || $this->casts === []) {
-            return $dates;
-        }
-
-        // Filter the ones that are casted as dates
-        $casts = array_filter($this->casts, function($k) {
-            return $k === 'date' || $k === 'datetime';
-        });
-
-        // Merge the key with our dates
-        $dates = array_merge($dates, array_keys($casts));
-
-        // Don't return duplicates
-        return array_unique(array_values($dates));
     }
 }
